app/requirements.txt
flask
gunicorn
tensorflow
pillow
requests
google-cloud-storage
fuzzywuzzy
numpy